requirements.txt
bentoml==1.4.18
fastapi==0.116.1
httpx==0.28.1
mlflow==2.17.0
mlflow_skinny==2.17.0
pandas==2.3.1
prometheus_client==0.21.1
prometheus_fastapi_instrumentator==7.1.0
pydantic==2.11.7
pytest==8.3.5
python-dotenv==1.1.1
starlette==0.47.2